ai-agent/requirements.txt
python-dotenv>=1.0.0
pydantic>=2.7
tqdm>=4.66
openai>=1.50.0
dashscope>=1.19.0
pypdf>=4.2.0
python-docx>=1.1.2
pillow>=10.4.0
numpy>=1.26
scikit-learn>=1.5
faiss-cpu>=1.8.0.post1
folium>=0.17.0
